api/requirements.txt
requests
flask>=2.0  
Werkzeug>=2.0
flask-cors
pillow==11.0.0
numpy==1.26.4